start adding top rated tv
