+++ lib/app_colors.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 
 abstract class AppColors {
-  static const kPrimaryColor = Color.fromRGBO(178, 156, 236, 1.0);
+  static const kPrimaryColorOld = Color.fromRGBO(178, 156, 236, 1.0);
   static const kSecondaryColor = Colors.purple;
 
-  static const kPrimaryColorNew = Color.fromRGBO(28, 26, 41, 1.0);
+  static const kPrimaryColor = Color.fromRGBO(28, 26, 41, 1.0);
   static const titleText = Color.fromRGBO(255, 255, 255, 1.0);
   static const genresText = Color.fromRGBO(222, 221, 223, 1.0);
   static const ratingText = Color.fromRGBO(119, 119, 119, 1.0);
