+++ lib/ui/widgets/movie_details/movie_details_widget.dart
@@ -55,7 +55,7 @@ class _MovieDetailsWidgetState extends State<MovieDetailsWidget> {
 
     return Scaffold(
       body: ColoredBox(
-        color: AppColors.kPrimaryColorNew,
+        color: AppColors.kPrimaryColor,
         child: ListView(
           children: [
             Column(
