+++ lib/ui/widgets/movie_top_rated/top_rated_movie_widget.dart
@@ -1,0 +1,1 @@
+// TODO: удалить или добавить?
