+++ lib/ui/widgets/tv_list/tv_list_widget.dart
@@ -1,130 +1,91 @@
 import 'package:comics_db_app/app_colors.dart';
 import 'package:comics_db_app/domain/api_client/api_client.dart';
-import 'package:comics_db_app/ui/widgets/tv_list/tv_list_model.dart';
+import 'package:comics_db_app/resources/resources.dart';
+import 'package:comics_db_app/ui/widgets/tv_popular/tv_popular_model.dart';
+import 'package:comics_db_app/ui/widgets/tv_top_rated/tv_top_rated_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
 
 class TvWidget extends StatelessWidget {
   const TvWidget({Key? key}) : super(key: key);
   @override
-  // TODO возможно сделать виджет закрытым
   Widget build(BuildContext context) => ChangeNotifierProvider(
-      create: (context) => TvListModel(), child: const TvListWidget());
+      create: (context) => TvPopularModel(), child: const TvListWidget());
 }
 
-
-class TvListWidget extends StatelessWidget {
+class TvListWidget extends StatefulWidget {
   const TvListWidget({Key? key}) : super(key: key);
+
+  @override
+  _TvListWidgetState createState() => _TvListWidgetState();
+}
+
+class _TvListWidgetState extends State<TvListWidget> {
   @override
   Widget build(BuildContext context) {
-    // final model = NotifierProvider.watch<TVListModel>(context);
-
-    // TODO не могу добавить пока не уберу кастомный провайдер
-    final model = Provider.of<TvListModel>(context, listen: true);
-    // final model = context.watch<TvListModel>();
-    if (model == null) return const SizedBox.shrink();
-    return Stack(
-      children: [
-        ListView.builder(
-            keyboardDismissBehavior: ScrollViewKeyboardDismissBehavior.onDrag,
-            padding: const EdgeInsets.only(top: 70.0),
-            itemCount: model.tvs.length,
-            itemExtent: 165,
-            itemBuilder: (BuildContext context, int index) {
-              model.showedTVAtIndex(index);
-              final tv = model.tvs[index];
-              final posterPath = tv.posterPath;
-              return Padding(
-                padding: const EdgeInsets.symmetric(
-                    horizontal: 16.0, vertical: 10.0),
-                child: Stack(
-                  children: [
-                    Container(
-                      decoration: BoxDecoration(
-                          color: Colors.white,
-                          border:
-                              Border.all(color: Colors.black.withOpacity(0.2)),
-                          borderRadius:
-                              const BorderRadius.all(Radius.circular(20.0)),
-                          boxShadow: [
-                            BoxShadow(
-                              color: Colors.black.withOpacity(0.1),
-                              blurRadius: 8,
-                              offset: const Offset(0, 2),
-                            )
-                          ]),
-                      clipBehavior: Clip.hardEdge,
-                      child: Row(
-                        children: [
-                          posterPath != null ? Image.network(ApiClient.imageUrl(posterPath), width: 95) : const SizedBox.shrink(),
-                          const SizedBox(width: 15.0),
-                          Expanded(
-                            child: Column(
-                              crossAxisAlignment: CrossAxisAlignment.start,
-                              children: [
-                                const SizedBox(height: 20.0),
-                                Text(
-                                  tv.name,
-                                  maxLines: 1,
-                                  overflow: TextOverflow.ellipsis,
-                                  style: const TextStyle(
-                                      fontWeight: FontWeight.bold),
-                                ),
-                                const SizedBox(height: 5.0),
-                                Text(
-                                  model.stringFromDate(tv.firstAirDate),
-                                  maxLines: 1,
-                                  overflow: TextOverflow.ellipsis,
-                                  style: const TextStyle(color: Colors.grey),
-                                ),
-                                const SizedBox(height: 20.0),
-                                Text(
-                                  tv.overview,
-                                  maxLines: 2,
-                                  overflow: TextOverflow.ellipsis,
-                                ),
-                              ],
-                            ),
-                          ),
-                          const SizedBox(width: 10.0),
-                        ],
-                      ),
+    // final topRatedTvModel = context.watch<TopRatedTvModel>();
+    return Scaffold(
+      appBar: AppBar(
+        titleSpacing: 0.0,
+        title: Padding(
+          padding: const EdgeInsets.symmetric(horizontal: 8.0),
+          child: Row(
+            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+            children: [
+              Row(
+                children: [
+                  Image.asset(AppImages.movieAppBarLogo),
+                ],
+              ),
+              Row(
+                crossAxisAlignment: CrossAxisAlignment.center,
+                mainAxisAlignment: MainAxisAlignment.end,
+                children: [
+                  GestureDetector(
+                    onTap: () {},
+                    child: const Icon(Icons.search, color: AppColors.searchIcon, size: 30,),
+                  ),
+                  Padding(
+                      padding: const EdgeInsets.symmetric(horizontal: 8.0),
+                    child: GestureDetector(
+                      onTap: () {},
+                      child: const Icon(Icons.filter_alt_outlined, color: AppColors.ratingThumb, size: 30,),
                     ),
-                    Material(
-                      color: Colors.transparent,
-                      child: InkWell(
-                        borderRadius: BorderRadius.circular(20.0),
-                        onTap: () => model.onTVTap(context, index),
-                      ),
-                    ),
-                  ],
-                ),
-              );
-            }),
-        Padding(
-          padding: const EdgeInsets.all(10.0),
-          child: TextField(
-            onChanged: model.searchTV,
-            decoration: InputDecoration(
-              labelText: 'Поиск',
-              labelStyle: const TextStyle(
-                color: AppColors.kPrimaryColor,
-              ),
-              filled: true,
-              fillColor: Colors.white.withAlpha(235),
-              border: OutlineInputBorder(
-                borderRadius: BorderRadius.circular(10.0),
-              ),
-              focusedBorder: const OutlineInputBorder(
-                borderSide: BorderSide(color: AppColors.kPrimaryColor),
-              ),
-              enabledBorder: const OutlineInputBorder(
-                borderSide: BorderSide(color: AppColors.kPrimaryColor),
-              ),
-            ),
+                  ),
+                  GestureDetector(
+                    onTap: () {},
+                    child: const Icon(Icons.menu, color: AppColors.ratingThumb, size: 30,),
+                  ),
+                ],
+              )
+            ],
           ),
+          ),
+        backgroundColor: AppColors.kPrimaryColor,
         ),
-      ],
-    );
+      body: ColoredBox(
+        color: AppColors.kPrimaryColor,
+        child: ListView(
+          children: [
+            Column(
+              children: [
+                const Padding(
+                    padding: EdgeInsets.only(top: 20.0, left: 20.0, bottom: 20.0),
+                  child: SizedBox(
+                    height: 180,
+                    child: TopRatedTvWidget(),
+                  ),
+                ),
+              ],
+            ),
+          ],
+        ),
+
+
+      ),
+      );
   }
 }
+
+
+
