+++ lib/ui/widgets/tv_popular/tv_popular_widget.dart
@@ -1,0 +1,118 @@
+import 'package:comics_db_app/app_colors.dart';
+import 'package:comics_db_app/domain/api_client/api_client.dart';
+import 'package:comics_db_app/ui/widgets/tv_popular/tv_popular_model.dart';
+import 'package:flutter/material.dart';
+import 'package:provider/provider.dart';
+
+class TvPopularWidget extends StatelessWidget {
+  const TvPopularWidget({Key? key}) : super(key: key);
+  @override
+  Widget build(BuildContext context) {
+    final model = Provider.of<TvPopularModel>(context, listen: true);
+    // final model = context.watch<TvListModel>();
+    if (model == null) return const SizedBox.shrink();
+    return Stack(
+      children: [
+        ListView.builder(
+            keyboardDismissBehavior: ScrollViewKeyboardDismissBehavior.onDrag,
+            padding: const EdgeInsets.only(top: 70.0),
+            itemCount: model.tvs.length,
+            itemExtent: 165,
+            itemBuilder: (BuildContext context, int index) {
+              model.showedTVAtIndex(index);
+              final tv = model.tvs[index];
+              final posterPath = tv.posterPath;
+              return Padding(
+                padding: const EdgeInsets.symmetric(
+                    horizontal: 16.0, vertical: 10.0),
+                child: Stack(
+                  children: [
+                    Container(
+                      decoration: BoxDecoration(
+                          color: Colors.white,
+                          border:
+                          Border.all(color: Colors.black.withOpacity(0.2)),
+                          borderRadius:
+                          const BorderRadius.all(Radius.circular(20.0)),
+                          boxShadow: [
+                            BoxShadow(
+                              color: Colors.black.withOpacity(0.1),
+                              blurRadius: 8,
+                              offset: const Offset(0, 2),
+                            )
+                          ]),
+                      clipBehavior: Clip.hardEdge,
+                      child: Row(
+                        children: [
+                          posterPath != null ? Image.network(ApiClient.imageUrl(posterPath), width: 95) : const SizedBox.shrink(),
+                          const SizedBox(width: 15.0),
+                          Expanded(
+                            child: Column(
+                              crossAxisAlignment: CrossAxisAlignment.start,
+                              children: [
+                                const SizedBox(height: 20.0),
+                                Text(
+                                  tv.name,
+                                  maxLines: 1,
+                                  overflow: TextOverflow.ellipsis,
+                                  style: const TextStyle(
+                                      fontWeight: FontWeight.bold),
+                                ),
+                                const SizedBox(height: 5.0),
+                                Text(
+                                  model.stringFromDate(tv.firstAirDate),
+                                  maxLines: 1,
+                                  overflow: TextOverflow.ellipsis,
+                                  style: const TextStyle(color: Colors.grey),
+                                ),
+                                const SizedBox(height: 20.0),
+                                Text(
+                                  tv.overview,
+                                  maxLines: 2,
+                                  overflow: TextOverflow.ellipsis,
+                                ),
+                              ],
+                            ),
+                          ),
+                          const SizedBox(width: 10.0),
+                        ],
+                      ),
+                    ),
+                    Material(
+                      color: Colors.transparent,
+                      child: InkWell(
+                        borderRadius: BorderRadius.circular(20.0),
+                        onTap: () => model.onTVTap(context, index),
+                      ),
+                    ),
+                  ],
+                ),
+              );
+            }),
+        Padding(
+          padding: const EdgeInsets.all(10.0),
+          child: TextField(
+            onChanged: model.searchTV,
+            decoration: InputDecoration(
+              labelText: 'Поиск',
+              labelStyle: const TextStyle(
+                color: AppColors.kPrimaryColorOld,
+              ),
+              filled: true,
+              fillColor: Colors.white.withAlpha(235),
+              border: OutlineInputBorder(
+                borderRadius: BorderRadius.circular(10.0),
+              ),
+              focusedBorder: const OutlineInputBorder(
+                borderSide: BorderSide(color: AppColors.kPrimaryColorOld),
+              ),
+              enabledBorder: const OutlineInputBorder(
+                borderSide: BorderSide(color: AppColors.kPrimaryColorOld),
+              ),
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+}
