+++ lib/ui/widgets/tv_top_rated/tv_top_rated_widget.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class TopRatedTvWidget extends StatelessWidget {
+  const TopRatedTvWidget({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container();
+  }
+}
